models interface shifted from dataAccess to movie.models.ts
