+++ FE/angular/src/app/components/movie/data-access/data-access.service.ts
@@ -2,28 +2,8 @@ import { Injectable } from '@angular/core';
 import { Observable, map } from 'rxjs';
 import { Constants } from 'src/app/constants/constants';
 import { HttpService } from 'src/app/services/http.service';
-interface Movie {
-  imdbID: string;
-  Poster: string;
-  Title: string;
-  Type: string;
-  Year: string | number;
-}
+import { MovieComplete, MovieDetails } from '../movie.models';
 
-interface MovieDetails extends Movie {
-  Actors: string;
-  Director: string;
-  Genre: string;
-  Plot: string;
-  Rated: string;
-  Released: string;
-  Runtime: string;
-  Writer: string;
-}
-
-export interface MovieComplete extends MovieDetails {
-  Year: number;
-}
 @Injectable({
   providedIn: 'root'
 })
